+++ Client/src/pages/special-offer.aspx.js
@@ -8,6 +8,8 @@ import ImageGQL from "../components/image-gql";
 import "../styles/avs-special-offer.less";
 import PageContentWrapper from '../components/page-content-wrapper';
 import Button from '../components/button';
+import Logo from '../images/common/logo.svg';
+import { Link as GatsbyLink } from 'gatsby';
 
 import ContentRowItem from '../components/content-row-item';
 
@@ -17,6 +19,20 @@ import customersIcon from '../images/avs-special-offer/customers-icon.png'
 import marketIcon from '../images/avs-special-offer/market-icon.png'
 import updatesIcon from '../images/avs-special-offer/updates-icon.png'
 
+
+const LogoWrapper = styled.div`
+    display: table;
+    position: relative;
+    width: 1120px;
+    margin: auto;
+    top: 12px;
+    left: 16px;
+    &>img{
+        width: 69px;
+        height: 60px;
+        margin: auto;
+    }
+`;
 
 const RedButton = styled(Button) `
     min-width: 155px;
@@ -36,10 +52,12 @@ const SaleWrapper = styled.div`
     }
 `;
 
+
 class SpecialOffer extends React.PureComponent {
 render(){
     return (
       <Layout 
+        headerIsDisabled={true}
         className="special-offer" 
         pageContext={this.props.pageContext} 
         t={this.props.t}
@@ -47,21 +65,27 @@ render(){
         metaDescription=""
         metaKeywords=""
       >
-
-<div className="headerSpecialOffer">
-    <div className="line-mobile"></div>
+        <link href="https://fonts.googleapis.com/css?family=Montserrat:100,100i,200,200i,300,400&display=swap" rel="stylesheet"></link>
+        <div className="header-specail-offer">
+            <LogoWrapper href="/">
+              <GatsbyLink className="headerLogo" to="/">
+                  <img src={Logo} alt="avs4you logo"/>
+                </GatsbyLink>
+            </LogoWrapper>
+        </div>
+<div className="headerSpecialOfferSummer">
       <div className="content-header">
       <div className="img-header">
-      <div className="snowflakes-header"></div>
-      <div className="multimedia-header"></div>
+      <div className="back-summer-header"></div>
+      <div className="multimedia-header-summer"></div>
       </div>
-      <div className="text-header">
-      <div className="avs-logo logo-offer"></div>
+      <div className="text-header text-header-summer">
+      <div className="avs-logo logo-summer-offer"></div>
         <Text fontSize={34} fontWeight={700} color="#fff" className="textSpecial">Special Offer</Text>
           <div className="text-grab">
           <Text fontSize={64} fontWeight={700} color="#fff" className="bodyGrab">Grab</Text>
             <SaleWrapper className="bodyOff">
-            <Text fontSize={130} fontWeight={700} color="#ffffff" className="precent-off">70</Text> 
+            <Text fontSize={130} fontWeight={700} fontFamily={'Montserrat'} color="#ffffff" className="precent-off">70</Text> 
               <SaleWrapper className="precent">
                 <Text fontSize={46} fontWeight={700} color="#ffffff" className="text-off">%</Text> 
                 <Text fontSize={46} fontWeight={700} color="#ffffff" className="letter-off">ff</Text>
@@ -69,39 +93,35 @@ render(){
             </SaleWrapper> 
             <div className="text-avs">
               <Text fontSize={24} fontWeight={700} color="#fff" className="text-header-avs">on</Text>
-              <Text fontSize={46} fontWeight={700} className="textHeader">AVS4YOU</Text>
+              <Text fontSize={46} fontWeight={700} className="textHeader-summer">AVS4YOU</Text>
             </div>
-            <Text fontSize={46} fontWeight={700} color="#fff" className="textMultiHeader">Multimedia Suite</Text>
+            <Text fontSize={46} fontWeight={700} color="#fff" className="textMultiHeader-summer">Multimedia Suite</Text>
             <div className="button-block">
-              <Button className="headerButton button-mobile" 
-                      secondaryText={this.props.secondaryText && this.props.secondaryText} 
-                      textTransform="uppercase">
-                      <Link to="/installed-pack.aspx" target="_blank" rel="noreferrer noopener" className="style-button"> download now </Link>
-              </Button>
               <div className="blockButton"></div>
-              <RedButton className="headerButton" 
+              <RedButton className="headerButtonSummer" 
                       secondaryText={this.props.secondaryText && this.props.secondaryText} 
                       textTransform="uppercase">
                       <Link to="/register.aspx" target="_blank" rel="noreferrer noopener" className="style-button"> buy now </Link>
               </RedButton>
             </div>
-            <Text fontSize={16} fontWeight={400} color="#fff" className="mainHeaderComment">The offer is valid till December 26, 2021</Text>
+            <Text fontSize={16} fontWeight={400} color="#fff" className="mainHeaderComment">The offer is valid during 24 hours</Text>
       </div>
     </div>
+    <div className="multimedia-header-summer-mobile"></div>
   </div>
 </div>
           <div className="bodyWrapper">
             <Text fontSize={46} fontWeight={800} className="bodyHeader" as="h1">AVS4YOU Full Pack</Text>
-            <Text fontSize={36} fontWeight={700} className="mainBodyTitle">Get 70% OFF on AVS4YOU Full Pack <span className="text-red">$59</span>  <span className="text-line-through">$199</span></Text>
+            <Text fontSize={36} fontWeight={700} className="mainBodyTitle">Get 70% OFF on AVS4YOU Full Pack <span className="text-red">$79</span> <span className="text-line-through">$279</span></Text>
             <Text fontSize={24} fontWeight={400} className="mainBodyComment">A complete set of programs for working with video files, slideshows and soundtracks. Manage your video/audio collections, improve their quality and create stunning video projects!</Text>
             <div className="special-buttons">
-              <Button className="specialButton button-mobile" 
+              <Button className="specialButtonSummer button-mobile" 
                       secondaryText={this.props.secondaryText && this.props.secondaryText} 
                       textTransform="uppercase">
                       <Link to="/installed-pack.aspx" target="_blank" rel="noreferrer noopener" className="style-button"> download now </Link>
               </Button>
               <div className="blockButton"></div>
-              <RedButton className="specialButton" 
+              <RedButton className="specialButtonSummer" 
                       secondaryText={this.props.secondaryText && this.props.secondaryText} 
                       textTransform="uppercase">
                       <Link to="/register.aspx" target="_blank" rel="noreferrer noopener" className="style-button"> buy now </Link>
@@ -169,31 +189,17 @@ render(){
                           <Text className="ListItem" color="#555555" fontWeight={400} fontSize={19}>Create your own audiobooks</Text>
                           <Text className="ListItem" color="#555555" fontWeight={400} fontSize={19}>Make customized ringtones for iPhone</Text>
                       </ContentRowItem>
-
-                      <ContentRowItem 
-                          disableBG={true}
-                          imgLeft={true}
-                          imageName={"remaker-special-offer.png"}
-                          headerText={this.props.t("AVS Video ReMaker")}
-                          translateButtonLink="https://downloads.avs4you.com/distributives/AVSVideoReMaker.exe" 
-                      >
-                        <Text  className="ListTitle" color="#555555" fontWeight={600} fontSize={24}>A fast video editor to make edits without changing the original video file format.</Text>
-                          <Text className="ListItem" color="#555555" fontWeight={400} fontSize={19}>Cut and delete unnecessary parts</Text>
-                          <Text className="ListItem" color="#555555" fontWeight={400} fontSize={19}>Change scene sequence in all popular video formats</Text>
-                          <Text className="ListItem" color="#555555" fontWeight={400} fontSize={19}>Transfer video from camcorders and video cameras to hard drive</Text>
-                          <Text className="ListItem" color="#555555" fontWeight={400} fontSize={19}>Create home video DVD or Blu-ray discs</Text>
-                      </ContentRowItem>
                   </PageContentWrapper>
 
                   <div className="why-choose">
                     <Text fontSize={46} fontWeight={800} className="bodyHeader" as="h1">Why Choose AVS4YOU</Text>
                     <div className="block-choose">
                       <div className="list-choose">
-                        <div className="choose-item"><img className="icon-mobile" src={secureIcon}/><Text className="awards-text" color="#555555" fontSize={24} fontWeight={600}>{this.props.t("100% Secure")}</Text></div>
-                        <div className="choose-item"><img className="icon-mobile" src={supportIcon}/><Text className="awards-text" color="#555555" fontSize={24} fontWeight={600}>{this.props.t("Free support")}</Text></div>
-                        <div className="choose-item"><img className="icon-mobile" src={updatesIcon}/><Text className="awards-text" color="#555555" fontSize={24} fontWeight={600}>{this.props.t("Free updates")}</Text></div>
-                        <div className="choose-item"><img className="icon-mobile" src={marketIcon}/><Text className="awards-text" color="#555555" fontSize={24} fontWeight={600}>{this.props.t("15+ years on the market")}</Text></div>
-                        <div className="choose-item"><img className="icon-mobile" src={customersIcon}/><Text className="awards-text" color="#555555" fontSize={24} fontWeight={600}>{this.props.t("20M happy customers")}</Text></div>
+                        <div className="choose-item"><img className="icon-mobile" src={secureIcon}/><Text className="awards-text" color="#555555" fontSize={24} fontWeight={600}>100% Secure</Text></div>
+                        <div className="choose-item"><img className="icon-mobile" src={supportIcon}/><Text className="awards-text" color="#555555" fontSize={24} fontWeight={600}>Free support</Text></div>
+                        <div className="choose-item"><img className="icon-mobile" src={updatesIcon}/><Text className="awards-text" color="#555555" fontSize={24} fontWeight={600}>Free updates</Text></div>
+                        <div className="choose-item"><img className="icon-mobile" src={marketIcon}/><Text className="awards-text" color="#555555" fontSize={24} fontWeight={600}>15+ years on the market</Text></div>
+                        <div className="choose-item"><img className="icon-mobile" src={customersIcon}/><Text className="awards-text" color="#555555" fontSize={24} fontWeight={600}>20M happy customers</Text></div>
                       </div>
                       </div>  
                   </div>