+++ Client/Changelog.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 1.0.32
+
+### New
+* Added page `/special-offer.aspx`
+
+### Fixes
+* Fixed version, date, file size
+* Fixed price
+
+## 1.0.31
+
+### Fixes
+* Changed links on `register.aspx`
+* Changed styles on `/avs-video-editor.aspx` 
+
 ## 1.0.30
 
 ### New